requirements.txt
beautifulsoup4
biopython
html5lib
lxml
mechanicalsoup
numpy
pandas
requests
saintBioutils
sqlalchemy>=1.4.20
tqdm
